Moved ConverterBase out of coordinates/base.py
