+++ package/CHANGELOG
@@ -13,7 +13,7 @@ The rules for this file:
   * release numbers follow "Semantic Versioning" http://semver.org
 
 ------------------------------------------------------------------------------
-??/??/?? IAlibay, hmacdope, pillose, jaclark5
+??/??/?? IAlibay, hmacdope, pillose, jaclark5, ianmkenney
 
  * 2.7.0
 
@@ -27,6 +27,7 @@ Changes
     (Issue #4247)
   * Cython DEF statments have been replaced with compile time integer constants
     as DEF is deprecated as of Cython>=3.0 (Issue #4237, PR #4246)
+  * ConverterBase class moved from coordinates/base.py to converters/base.py (Issue #3404)
 
 Deprecations
