+++ package/MDAnalysis/coordinates/base.py
@@ -101,17 +101,6 @@ file.
    :members:
    :inherited-members:
 
-
-Converters
-----------
-
-Converters output information to other libraries.
-
-.. autoclass:: ConverterBase
-   :members:
-   :inherited-members:
-
-
 Helper classes
 --------------
 
@@ -134,7 +123,6 @@ from .. import (
     _READERS, _READER_HINTS,
     _SINGLEFRAME_WRITERS,
     _MULTIFRAME_WRITERS,
-    _CONVERTERS
 )
 from .. import units
 from ..auxiliary.base import AuxReader
@@ -1782,31 +1770,3 @@ def range_length(start, stop, step):
     else:
         # The range is empty.
         return 0
-
-
-class _Convertermeta(type):
-    # Auto register upon class creation
-    def __init__(cls, name, bases, classdict):
-        type.__init__(type, name, bases, classdict)
-        try:
-            fmt = asiterable(classdict['lib'])
-        except KeyError:
-            pass
-        else:
-            for f in fmt:
-                f = f.upper()
-                _CONVERTERS[f] = cls
-
-class ConverterBase(IOBase, metaclass=_Convertermeta):
-    """Base class for converting to other libraries.
-
-    See Also
-    --------
-    :mod:`MDAnalysis.converters`
-    """
-
-    def __repr__(self):
-        return "<{cls}>".format(cls=self.__class__.__name__)
-
-    def convert(self, obj):
-        raise NotImplementedError
